refactor: moved the RSS feed widget into a separate package
